requirements.txt
Flask
flask-bcrypt
flask-login
pandas
scikit-learn
joblib
geopy
requests
python-dotenv
gunicorn
